\name{rabbitandchicken-package}
\alias{rabbitandchicken-package}
\alias{rabbitandchicken}
\docType{package}
\title{
  return rabbits
}
\description{
  A more detailed description of what the package does. A length
  of about one to five lines is recommended.
}
\details{
  This section should provide a more detailed overview of how to use the
  package, including the most important functions.
}
\author{
Notplayingcute.

Maintainer: Notplayingcute <1613104@nankai.edu.cn>
}
\references{
  This optional section can contain literature or other references for
  background information.
}
\keyword{ package }
\seealso{
  Optional links to other man pages
}
\examples{
  \dontrun{
     ## Optional simple examples of the most important functions
     ## These can be in \dontrun{} and \donttest{} blocks.   
  }
}
